Add sdf command generator file
